Prevent mouse wheel scroll on entry edit username field
* Reported by shawnkhu via Matrix
