+++ src/gui/entry/EditEntryWidgetMain.ui
@@ -140,6 +140,9 @@
     </item>
     <item row="1" column="1">
      <widget class="QComboBox" name="usernameComboBox">
+      <property name="focusPolicy">
+       <enum>Qt::StrongFocus</enum>
+      </property>
       <property name="accessibleName">
        <string>Username field</string>
       </property>
